State time is now derived from animation clip length, rather than hard coded.

+++ Assets/Scripts/PlayerState.cs
@@ -33,7 +33,7 @@ public class PlayerState : MonoBehaviour {
 		}
 	}
 
-	public void SetStateTimer (Moves move, int frames) {
+	public void SetStateTimer (Moves move, float frames) {
 		SetMoveState(move);
 		PlayerMoved(this, new EvolutionEventArgs(move));
 		stateTimer.Interval = (frames * 1000) / 60;
